+++ src/app/protect-account/page.tsx
@@ -17,20 +17,12 @@ import LoadingModal from "@/components/loading-modal";
 import { sepolia } from "wagmi/chains";
 import { BaseForm } from "@/components/base-form";
 import { useGuardians } from "@/hooks/useGuardians";
-import { SrmAddress } from "@/hooks/use-social-recovery-module";
-import { SocialRecoveryModuleGracePeriodSelector } from "abstractionkit";
 import { getEtherscanAddressLink } from "@/utils/get-etherscan-link";
+import { delayPeriodMap } from "@/utils/delay-period";
 
 const totalSteps = 4;
 
 const isBrowser = typeof window !== "undefined";
-
-export const delayPeriodMap: Record<string, SrmAddress> = {
-  [1]: SocialRecoveryModuleGracePeriodSelector.After3Minutes,
-  [3]: SocialRecoveryModuleGracePeriodSelector.After3Days,
-  [7]: SocialRecoveryModuleGracePeriodSelector.After7Days,
-  [14]: SocialRecoveryModuleGracePeriodSelector.After14Days,
-};
 
 export default function ProtectAccount() {
   const [currentStep, setCurrentStep] = useState(1);
